Completed NonLin Residual
Finished code fully implementing the non-linear residual calculation described. All jacobians and what not are calculated. The results were also tested and shown to be correct, assuming that the underlying constituitive law is correctly implemented. This has still not been determined. All code is stable under testing except for the tests on Assignment 9.